+++ src/services/mediapire/mediapire.ts
@@ -1,8 +1,16 @@
+import api from "../../api/api";
+
 const localStorageKey = "mediapire-ui-manager";
+
+type MediapireSettings = {
+  fileTypes: string[];
+};
+
 interface MediapireService {
   getManagerConfig: () => MediapireManageConfig | null;
   saveManagerConfig: (config: MediapireManageConfig) => void;
   deleteManagerConfig: () => void;
+  getSettings: () => Promise<MediapireSettings>;
 }
 
 const getManagerConfig = (): MediapireManageConfig | null => {
@@ -21,8 +29,13 @@ const saveManagerConfig = (config: MediapireManageConfig): void =>
 const deleteManagerConfig = (): void =>
   localStorage.removeItem(localStorageKey);
 
+const getSettings = (): Promise<MediapireSettings> => {
+  return api.get("/api/v1/settings").then((r) => r.json());
+};
+
 export const mediapireService: MediapireService = {
   getManagerConfig,
   saveManagerConfig,
   deleteManagerConfig,
+  getSettings,
 };
