+++ src/utils/constants.ts
@@ -1,4 +1,6 @@
-export const routeLibrary = "/library";
+export const libraryBasePath = "/library";
+export const LIBRARY_MEDIA_ID_PARAM = ":mediaType?";
+export const routeMediaLibrary = `${libraryBasePath}/${LIBRARY_MEDIA_ID_PARAM}`;
 export const routeSetup = "/setup";
 export const routeError = "/error";
 export const routeDownloadStatus = "/downloads/:id";
